PrintingConfig cleanup
